+++ lib/types/date.dart
@@ -56,11 +56,11 @@ class JarDate extends JarSchema<DateTime, JarDate> {
     });
   }
 
-  JarDate custom(String? Function(DateTime? value) validator,
+  JarDate custom(String? Function(DateTime? value, [Map<String, dynamic>? allValues]) validator,
       [String? message]) {
     return addValidator((value) {
       if (value == null) return null;
-      final validationResult = validator(value);
+      final validationResult = validator(value, getAllValues());
       return validationResult ?? null;
     });
   }